+++ src/peer/media/sender/mod.rs
@@ -32,13 +32,13 @@ pub use self::component::{Component, State};
 pub enum CreateError {
     /// [`Sender`] cannot be disabled because it's marked as `required`.
     #[display(
-        "MediaExchangeState of Sender cannot transit to \
-                     disabled state, because this Sender is required."
+        "`MediaExchangeState` of `Sender` cannot transit to \
+         disabled state, because this `Sender` is required"
     )]
     CannotDisableRequiredSender,
 
     /// Could not find a [`platform::Transceiver`] by `mid`.
-    #[display("Unable to find Transceiver with mid: {}", _0)]
+    #[display("Unable to find Transceiver with mid: {_0}")]
     TransceiverNotFound(String),
 }
 
@@ -47,7 +47,7 @@ pub enum CreateError {
 /// [1]: https://w3.org/TR/webrtc#dom-rtcrtpsender-replacetrack
 #[derive(Caused, Clone, Debug, Display, From)]
 #[cause(error = platform::Error)]
-#[display("MediaManagerHandle is in detached state")]
+#[display("`MediaManagerHandle` is in detached state")]
 pub struct InsertTrackError(platform::Error);
 
 /// Representation of a [`local::Track`] that is being sent to some remote peer.