+++ tests/peer/mod.rs
@@ -1397,11 +1397,8 @@ async fn new_remote_track() {
                 has_audio: audio_tx_enabled && audio_rx_enabled,
                 has_video: video_tx_enabled && video_rx_enabled,
             },
-            "{} {} {} {}",
-            audio_tx_enabled,
-            video_tx_enabled,
-            audio_rx_enabled,
-            video_rx_enabled,
+            "{audio_tx_enabled} {video_tx_enabled} \
+             {audio_rx_enabled} {video_rx_enabled}",
         );
     }
 }