+++ tests/room/room.rs
@@ -162,7 +162,7 @@ async fn error_get_local_stream_on_new_peer() {
         cb_assert_eq!(
             &err.message(),
             "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-             Error: error_get_local_stream_on_new_peer"
+             Error: error_get_local_stream_on_new_peer",
         );
         assert_eq!(&cause.message(), "error_get_local_stream_on_new_peer");
         assert!(&err.trace().contains("at src"));
@@ -216,7 +216,7 @@ async fn error_join_room_without_on_failed_stream_callback() {
 
     assert_eq!(
         err.message(),
-        "`Room.on_failed_local_media()` callback isn't set.",
+        "`Room.on_failed_local_media()` callback isn't set",
     );
     assert!(!err.trace().is_empty());
 }
@@ -247,7 +247,7 @@ async fn error_join_room_without_on_connection_loss_callback() {
 
     assert_eq!(
         err.message(),
-        "`Room.on_connection_loss()` callback isn't set.",
+        "`Room.on_connection_loss()` callback isn't set",
     );
     assert!(!err.trace().is_empty());
 }
@@ -1432,9 +1432,8 @@ mod rpc_close_reason_on_room_drop {
         assert_eq!(
             close_reason,
             ClientDisconnect::RoomUnexpectedlyDropped,
-            "Room sets RPC close reason '{:?} instead of \
-             'RoomUnexpectedlyDropped'.",
-            close_reason,
+            "`Room` sets RPC close reason '{close_reason:?} instead of \
+             'RoomUnexpectedlyDropped'",
         )
     }
 
@@ -1462,8 +1461,8 @@ mod rpc_close_reason_on_room_drop {
         assert_eq!(
             close_reason,
             ClientDisconnect::RoomClosed,
-            "Room sets RPC close reason '{:?}' instead of 'RoomClosed.",
-            close_reason,
+            "`Room` sets RPC close reason '{close_reason:?}' instead of \
+             'RoomClosed",
         );
     }
 }
@@ -2195,8 +2194,7 @@ async fn set_media_state_return_media_error() {
         err.message(),
         format!(
             "Failed to get local tracks: MediaDevices.getUserMedia() failed: \
-             Error: {}",
-            ERROR_MSG
+             Error: {ERROR_MSG}",
         )
     );
 
@@ -2513,8 +2511,8 @@ mod set_local_media_settings {
                 jsval_cast(err, "MediaStateTransitionException")?;
             cb_assert_eq!(
                 err.message(),
-                "MediaExchangeState of Sender cannot transit to \
-                 disabled state, because this Sender is required."
+                "`MediaExchangeState` of `Sender` cannot transit to disabled \
+                 state, because this `Sender` is required",
             );
         });
         room_handle.on_failed_local_media(cb.into()).unwrap();
@@ -2561,7 +2559,7 @@ mod set_local_media_settings {
                 jsval_cast(err, "MediaStateTransitionException")?;
             cb_assert_eq!(
                 &err.message(),
-                "provided multiple device video MediaStreamTracks"
+                "provided multiple device video MediaStreamTracks",
             );
         });
         let (audio_track, video_track) = get_test_required_tracks();
@@ -2641,12 +2639,10 @@ mod set_local_media_settings {
                 .await
                 .is_ok(),
                 is_should_be_ok,
-                "audio_required: {}; add_audio: {}; video_required: {}; \
-                 add_video: {}",
-                audio_required,
-                add_audio,
-                video_required,
-                add_video,
+                "audio_required: {audio_required}; \
+                 add_audio: {add_audio}; \
+                 video_required: {video_required}; \
+                 add_video: {add_video}",
             );
 
             timeout(1000, test_rx)