+++ proto/client-api/CHANGELOG.md
@@ -6,6 +6,19 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.6.1] · unreleased
+[0.6.1]: /../../tree/medea-client-api-proto-0.6.1/proto/client-api
+
+### Upgraded
+
+- Dependencies:
+    - [`derive-more`] to `1.0` ([#181]).
+
+[#181]: /../../pull/181
+
+
+
+
 ## [0.6.0] · 2024-08-05
 [0.6.0]: /../../tree/medea-client-api-proto-0.6.0/proto/client-api
 
@@ -220,5 +233,5 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
-
+[`derive-more`]: https://docs.rs/derive_more
 [Semantic Versioning 2.0.0]: https://semver.org